requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
requests==2.28.2
sqlparse==0.4.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.15
